WIP, TODO: change to encoding/json
